Moved Utils.getDefinedOwnProps to common
