+++ src/m0_facilitator/repositories/AuxiliaryChainRepository.ts
@@ -21,7 +21,7 @@ import { DataTypes, InitOptions, Model } from 'sequelize';
 
 import AuxiliaryChain from '../models/AuxiliaryChain';
 import Subject from '../../common/observer/Subject';
-import Utils from '../Utils';
+import Utils from '../../common/Utils';
 
 /**
  * An interface, that represents a row from a auxiliary_chains table.
