+++ src/m0_facilitator/repositories/GatewayRepository.ts
@@ -21,7 +21,7 @@ import { DataTypes, InitOptions, Model } from 'sequelize';
 
 import Gateway from '../models/Gateway';
 import Subject from '../../common/observer/Subject';
-import Utils from '../Utils';
+import Utils from '../../common/Utils';
 import { MAX_VALUE } from '../Constants';
 
 /**
