+++ src/m1_facilitator/repositories/DepositIntentRepository.ts
@@ -22,7 +22,7 @@ import { DataTypes, InitOptions, Model } from 'sequelize';
 import { MAX_VALUE } from '../../m0_facilitator/Constants';
 import DepositIntent from '../models/DepositIntent';
 import Subject from '../../common/observer/Subject';
-import Utils from '../../m0_facilitator/Utils';
+import Utils from '../../common/Utils';
 
 /**
  * An interface, that represents a row from a deposit_intents table.
