+++ src/m1_facilitator/repositories/ERC20GatewayTokenPairRepository.ts
@@ -16,7 +16,7 @@ import { DataTypes, Model, InitOptions } from 'sequelize';
 
 import ERC20GatewayTokenPair from '../models/ERC20GatewayTokenPair';
 import Subject from '../../common/observer/Subject';
-import Utils from '../../m0_facilitator/Utils';
+import Utils from '../../common/Utils';
 
 import assert = require('assert');
