+++ src/m1_facilitator/repositories/GatewayRepository.ts
@@ -19,7 +19,7 @@ import assert from 'assert';
 import BigNumber from 'bignumber.js';
 import { DataTypes, InitOptions, Model } from 'sequelize';
 import Gateway, { GatewayType } from '../models/Gateway';
-import Utils from '../../m0_facilitator/Utils';
+import Utils from '../../common/Utils';
 import Subject from '../../common/observer/Subject';
 
 /**
