+++ src/m1_facilitator/repositories/WithdrawIntentRepository.ts
@@ -21,7 +21,7 @@ import { DataTypes, InitOptions, Model } from 'sequelize';
 
 import { MAX_VALUE } from '../../m0_facilitator/Constants';
 import Subject from '../../common/observer/Subject';
-import Utils from '../../m0_facilitator/Utils';
+import Utils from '../../common/Utils';
 import WithdrawIntent from '../models/WithdrawIntent';
 
 /**
